Added a test for duplicate xml:id in results.
When eXist-db returns results from queries, the results are put in an
exist:result element. Suppose that two elements in two valid XML documents
stored in eXist have the same xml:id. This is fine because the elements
are in different documents. Now if the a query is such that the two
elements are part of the results, the resulting XML won't be valid
because it will contain the same xml:id value twice. However, the XML
will be well-formed. The added test checks that eulexistdb is able to
return the well-formed results. (That is, it won't try to validate the
result and raise a validation error.)